+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
+
+## [Unreleased]
+
+### Changed
+
+- Moved `get_timezone_choices()` to `middleware.py`
+
 ## [0.0.2] - 2026-02-26
 
 ### Added